+++ app/services/worldpay_service.rb
@@ -14,9 +14,12 @@ class WorldpayService
     reference = parse_response(response)
 
     if reference.present?
+      worldpay_url_service = WorldpayUrlService.new(@transient_registration.reg_identifier, reference[:link])
+      url = worldpay_url_service.format_link
+
       {
         payment: new_payment_object(@order),
-        url: format_link(reference[:link])
+        url: url
       }
     else
       :error
@@ -98,47 +101,6 @@ class WorldpayService
 
   def new_payment_object(order)
     Payment.new_from_worldpay(order)
-  end
-
-  def format_link(url)
-    [url,
-     success_url,
-     pending_url,
-     failure_url,
-     cancel_url,
-     error_url].join
-  end
-
-  def success_url
-    ["&successURL=", success_path].join
-  end
-
-  def pending_url
-    ["&pendingURL=", failure_path].join
-  end
-
-  def failure_url
-    ["&failureURL=", failure_path].join
-  end
-
-  def cancel_url
-    ["&cancelURL=", failure_path].join
-  end
-
-  def error_url
-    ["&errorURL=", failure_path].join
-  end
-
-  def success_path
-    url = [Rails.configuration.wcrs_renewals_url,
-           Rails.application.routes.url_helpers.success_worldpay_forms_path(@transient_registration.reg_identifier)]
-    CGI.escape(url.join)
-  end
-
-  def failure_path
-    url = [Rails.configuration.wcrs_renewals_url,
-           Rails.application.routes.url_helpers.failure_worldpay_forms_path(@transient_registration.reg_identifier)]
-    CGI.escape(url.join)
   end
 
   # Validating WorldPay responses
